Refactor: MyCourseLocation/Tag View Style
