--- app/src/main/java/cmc/sole/android/MyCourse/MyCourseWriteLocationBottomFragment.kt
@@ -1,89 +1,0 @@
-package cmc.sole.android.MyCourse
-
-import android.app.Activity
-import android.app.Dialog
-import android.os.Bundle
-import android.util.DisplayMetrics
-import android.util.Log
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.activityViewModels
-import cmc.sole.android.MyCourse.Write.MyCourseWriteViewModel
-import cmc.sole.android.R
-import cmc.sole.android.Utils.RecyclerViewDecoration.RecyclerViewHorizontalDecoration
-import cmc.sole.android.Utils.RecyclerViewDecoration.RecyclerViewVerticalDecoration
-import cmc.sole.android.databinding.BottomFragmentMyCourseWriteLocationBinding
-import cmc.sole.android.databinding.BottomFragmentMyCourseWriteTagBinding
-import com.google.android.flexbox.FlexboxLayoutManager
-import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.google.android.material.bottomsheet.BottomSheetDialog
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-
-class MyCourseWriteLocationBottomFragment: BottomSheetDialogFragment() {
-
-    lateinit var binding: BottomFragmentMyCourseWriteLocationBinding
-    private lateinit var dialogFinishListener: OnLocationFragmentFinishListener
-
-    interface OnLocationFragmentFinishListener {
-        fun finish()
-    }
-
-    fun setOnFinishListener(listener: OnLocationFragmentFinishListener) {
-        dialogFinishListener = listener
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        dialogFinishListener.finish()
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        binding = BottomFragmentMyCourseWriteLocationBinding.inflate(inflater, container, false)
-
-        initAdapter()
-        initClickListener()
-
-        return binding.root
-    }
-
-    private fun initClickListener() {
-
-    }
-
-    private fun initAdapter() {
-
-    }
-
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        val dialog = super.onCreateDialog(savedInstanceState)
-        dialog.setOnShowListener {
-            val bottomSheetDialog = it as BottomSheetDialog
-            setupRatio(bottomSheetDialog)
-        }
-        return dialog
-    }
-
-    private fun setupRatio(bottomSheetDialog: BottomSheetDialog) {
-        val bottomSheet = bottomSheetDialog.findViewById<View>(com.google.android.material.R.id.design_bottom_sheet) as View
-        val behavior = BottomSheetBehavior.from<View>(bottomSheet)
-        val layoutParams = bottomSheet!!.layoutParams
-        layoutParams.height = getBottomSheetDialogDefaultHeight()
-        bottomSheet.layoutParams = layoutParams
-        behavior.state = BottomSheetBehavior.STATE_EXPANDED
-    }
-
-    private fun getBottomSheetDialogDefaultHeight(): Int {
-        return getWindowHeight() * 85 / 100
-    }
-
-    private fun getWindowHeight(): Int {
-        val displayMetrics = DisplayMetrics()
-        (context as Activity?)!!.windowManager.defaultDisplay.getMetrics(displayMetrics)
-        return displayMetrics.heightPixels
-    }
-}
