+++ app/src/main/res/layout/bottom_fragment_my_course_write_tag.xml
@@ -3,16 +3,48 @@
     android:layout_width="match_parent"
     android:layout_height="600dp"
     xmlns:tools="http://schemas.android.com/tools"
+    android:paddingTop="24dp"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@drawable/bottom_sheet_my_course_tag">
+
+    <TextView
+        android:id="@+id/my_course_bottom_title"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="취향"
+        android:textColor="@color/black"
+        android:textSize="16sp"
+        android:layout_marginStart="16dp"
+        android:fontFamily="@font/pretendard_semibold"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageView
+        android:layout_width="32dp"
+        android:layout_height="32dp"
+        android:src="@drawable/ic_cancel_x"
+        android:layout_marginEnd="16dp"
+        app:layout_constraintTop_toTopOf="@id/my_course_bottom_title"
+        app:layout_constraintBottom_toBottomOf="@id/my_course_bottom_title"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:tint="#404040" />
+
+    <View
+        android:id="@+id/my_course_bottom_line"
+        android:layout_width="match_parent"
+        android:layout_height="1dp"
+        android:background="#EDEDED"
+        android:layout_marginTop="20dp"
+        app:layout_constraintTop_toBottomOf="@id/my_course_bottom_title" />
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
         android:layout_height="0dp"
+        android:layout_marginTop="20dp"
         android:layout_marginBottom="20dp"
         android:scrollbars="none"
         android:overScrollMode="never"
-        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/my_course_bottom_line"
         app:layout_constraintBottom_toTopOf="@id/my_course_tag_ok_btn">
 
         <androidx.constraintlayout.widget.ConstraintLayout
@@ -99,13 +131,13 @@
         </androidx.constraintlayout.widget.ConstraintLayout>
 
     </androidx.core.widget.NestedScrollView>
-    
+
     <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/my_course_tag_ok_btn"
         android:layout_width="match_parent"
         android:layout_height="48dp"
         android:background="@drawable/default_button_o"
-        android:text="태그태그"
+        android:text="적용하기"
         android:textColor="@color/white"
         android:textSize="16sp"
         android:fontFamily="@font/pretendard_regular"
